Add new "Tests" project to ClipboardPlugin solution
This commit introduces a new project named "Tests" to the `ClipboardPlugin.sln` solution file. The project is associated with the GUID `{2150E333-8FDC-42A3-9474-1A3956D46DE8}` and is located in the "Tests" directory. The nested project references have been updated to include this new project, replacing the previous references for "ClipboardPlugin.ExpressionEngine.Tests" and "ClipboardPlugin.Tests" with the new GUIDs.